Added a preprocessor define to disable .dem file association when in debug mode.
The floppy drive icon is now loaded from a static resource, rather than retrieving it through the WIN32 API, which requires too much effort to disable reading from the floppy device. This stops the floppy drive from being accessed on start up.

Removed unnecessary and dodgy protocol 46 check to determine for CS demos to determine whether the demo was recorded with 1.5 or an early version of 1.6. The client DLL checksum is the sole method used to determine this now.